requirements.txt
google-cloud-storage>=2.10.0
pandas>=2.0.0 